requirements_bcat.txt
# core BCAT / PyTorch stack (CUDA 12 wheels work on Colab T4/L4/P4/V100/A100)
torch==2.5.1+cu121,<3.0
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121
--extra-index-url https://download.pytorch.org/whl/cu121

# scientific + ML
numpy==2.0.2
scipy==1.14.1
pandas==2.2.2
scikit-learn==1.5.2
h5py==3.12.1
einops==0.8.0
hydra-core==1.3.2
statsmodels==0.14.4
tensorly==0.9.0
tensorly-torch==0.5.0

# plotting / notebook
matplotlib==3.9.2
seaborn==0.13.2
jupyterlab==4.3.1
ipykernel==6.17.1
jupyter-client>=5.2.0,<8.0
ipywidgets==8.1.5
tqdm==4.67.0

# misc utilities
pyyaml==6.0.2
packaging==24.2
requests==2.32.3
protobuf==5.29.1
rich==13.9.4
wandb==0.18.7

# from the original pip block (already pip-ready)
absl-py==2.1.0
accelerate==1.1.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.4
aiosignal==1.3.1
astunparse==1.6.3
black==24.10.0
blobfile==3.0.0
configmypy==0.2.0
dadaptation==3.2
datasets==3.1.0
dill==0.3.8
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.6.1
gcsfs==2024.6.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
huggingface-hub>=0.28.1
iniconfig==2.0.0
jax==0.5.1
jaxlib==0.5.1
keras==3.6.0
libclang==18.1.1
lightning-utilities==0.11.8
lxml==5.3.0
markdown==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
ml-dtypes==0.4.1
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.4.2
neuraloperator==0.3.0
opt-einsum==3.4.0
optree==0.13.1
pathspec==0.12.1
pluggy==1.5.0
propcache==0.2.0
pyarrow==18.0.0
pyclean==3.0.0
pycryptodomex==3.21.0
pytest==8.3.3
pytest-mock==3.14.0
regex==2024.11.6
rotary-embedding-torch==0.8.6
ruamel-yaml==0.18.6
ruamel-yaml-clib==0.2.12
safetensors==0.4.5
sentencepiece==0.2.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tiktoken==0.8.0
tokenizers==0.20.3
torch-harmonics==0.6.5
torchao==0.6.1
torchdata==0.7.1
torchinfo==1.8.0
torchmetrics==1.6.0
torchtune==0.4.0
transformers==4.46.3
triton==3.1.0
wrapt==1.17.2
xformers==0.0.28.post3
xxhash==3.5.0
yarl==1.17.2
tabulate==0.9.0
torch-optimizer==0.3.0